==> scripts/python_operations/notebooks/2_compute_rmsd_and_correlation_on_merged_data.ipynb <==
In this notebook, RMSD and Correlation values are obtained

import pandas as pd
import numpy as np
import seaborn as sns

import geopandas as gpd
import matplotlib.pyplot as plt
import proplot as pplt

from scipy.stats import spearmanr

root = "/Users/dimeji/Documents/Research/2024/deforestation/final_project_version/data/tabular_data/"

forested_data = gpd.read_file(root+"merged_forested_data.geojson")
deforested_data = gpd.read_file(root+"merged_deforested_data.geojson")

# If df is a GeoDataFrame and the columns are as described, focus on the data columns:
data_columns = ['ESA', 'FROM_GLC', 'DW', 'ESRI', 'MFGFC', 'PALSAR']
forested_data_select = forested_data[data_columns]
deforested_data_select = deforested_data[data_columns]

# Calculate Spearman correlation matrix
corr_matrix_deforested = deforested_data_select.corr(method='spearman')

# Create a mask to zero out the upper triangle including the diagonal
mask = np.triu(np.ones_like(corr_matrix_deforested, dtype=bool))

# Apply the mask to the correlation matrix
triangular_corr_matrix_deforested = corr_matrix_deforested.where(~mask).iloc[1:,:-1]

plt.figure(figsize=(11, 9))  
ax = sns.heatmap(triangular_corr_matrix_deforested, annot=True, fmt=".2f", cmap='warmcool', cbar=True,
                      cbar_kws={'label': "Spearman's Correlation Coefficient"},
                      square=True, linewidths=0.2,linecolor='black',
                      annot_kws={'fontsize':25}, vmin=0, vmax=1)

# Customize the plot
plt.xticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed
plt.yticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed 
#plt.yticks(fontsize=12)

cbar = ax.collections[0].colorbar
cbar.set_label("Spearman's Correlation Coefficient", size=25)  # Set font size for the label
cbar.ax.yaxis.set_tick_params(labelsize=25)
plt.tight_layout()
plt.savefig('deforested_data_correlation.png', dpi=400)
plt.show()

# Calculate Spearman correlation matrix
corr_matrix_forested = forested_data_select.corr(method='spearman')

# Apply the mask to the correlation matrix
triangular_corr_matrix_forested = corr_matrix_forested.where(~mask).iloc[1:,:-1]

plt.figure(figsize=(11, 9))  
ax = sns.heatmap(triangular_corr_matrix_forested, annot=True, fmt=".2f", cmap='warmcool', cbar=True,
                      cbar_kws={'label': "Spearman's Correlation Coefficient"},
                      square=True, linewidths=0.2,linecolor='black',
                      annot_kws={'fontsize':25}, vmin=0, vmax=1)

# Customize the plot
plt.xticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed
plt.yticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed 
#plt.yticks(fontsize=12)

cbar = ax.collections[0].colorbar
cbar.set_label("Spearman's Correlation Coefficient", size=25)  # Set font size for the label
cbar.ax.yaxis.set_tick_params(labelsize=25)
plt.tight_layout()
plt.savefig('forested_data_correlation.png', dpi=400)

plt.show()

# RMSD

def calculate_rmse(series1, series2):
    
    
    mask = ~np.isnan(series1) & ~np.isnan(series2)
    return np.sqrt(np.mean(((series1[mask] / 1000000) - (series2[mask] / 1000000)) ** 2))

# Initialize an empty DataFrame for storing RMSE values
variables = ['ESA', 'FROM_GLC', 'DW', 'ESRI', 'MFGFC', 'PALSAR']
rmsd_matrix_forested = pd.DataFrame(index=variables, columns=variables)
rmsd_matrix_deforested = pd.DataFrame(index=variables, columns=variables)

# Compute RMSE for every pair of variables
for var1 in variables:
    for var2 in variables:
        rmsd_matrix_forested.loc[var1, var2] = calculate_rmse(
                                            forested_data_select[var1], 
                                            forested_data_select[var2]
                                            )
rmsd_matrix_forested = rmsd_matrix_forested.astype(float)
triangular_rmsd_matrix_forested  = rmsd_matrix_forested .where(~mask).iloc[1:,:-1]

# Compute RMSE for every pair of variables
for var1 in variables:
    for var2 in variables:
        rmsd_matrix_deforested.loc[var1, var2] = calculate_rmse(
                                            deforested_data_select[var1], 
                                            deforested_data_select[var2]
                                            )
rmsd_matrix_deforested = rmsd_matrix_deforested.astype(float)
triangular_rmsd_matrix_deforested  = rmsd_matrix_deforested .where(~mask).iloc[1:,:-1]

plt.figure(figsize=(11, 9))  
ax =sns.heatmap(
        triangular_rmsd_matrix_forested, 
        annot=True, 
        fmt=".2f", 
        cmap='coolwarm',
        cbar=True,
        linewidths=0.2,
        linecolor='black',
        annot_kws={'fontsize':25},
        cbar_kws={'label': 'RMSD in square kilometers (km²)'})

# Customize the plot
plt.xticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed
plt.yticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed 
#plt.yticks(fontsize=12)

cbar = ax.collections[0].colorbar
cbar.set_label('RMSD in square kilometers (km²)', size=25)  # Set font size for the label
cbar.ax.yaxis.set_tick_params(labelsize=25)
plt.tight_layout()
plt.savefig('forested_data_rmsd.png', dpi=400)

plt.show()

plt.figure(figsize=(11, 9))  
ax =sns.heatmap(
        triangular_rmsd_matrix_deforested, 
        annot=True, 
        fmt=".2f", 
        cmap='coolwarm',
        cbar=True,
        linewidths=0.2, 
        linecolor='black',
        annot_kws={'fontsize':25},
        cbar_kws={'label': 'RMSD in square kilometers (km²)'})

# Customize the plot
plt.xticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed
plt.yticks(rotation=45, ha='right', fontsize=25)  # Customize tick labels as needed 
#plt.yticks(fontsize=12)

cbar = ax.collections[0].colorbar
cbar.set_label('RMSD in square kilometers (km²)', size=25)  # Set font size for the label
cbar.ax.yaxis.set_tick_params(labelsize=25)
plt.tight_layout()
plt.savefig('deforested_data_rmsd.png', dpi=400)

plt.show()

deforested_data_select_km = deforested_data_select/1000000
forested_data_select_km = forested_data_select/1000000

g = sns.pairplot(deforested_data_select_km, plot_kws={'s': 10, 'color': 'blue'}, diag_kws={'color': 'blue'})
for ax in g.axes.flatten():
    plt.setp(ax.get_xticklabels(), fontsize=12, fontname='Arial')
    plt.setp(ax.get_yticklabels(), fontsize=12, fontname='Arial')
    ax.set_xlabel(ax.get_xlabel(), fontsize=23, fontname='Arial')
    ax.set_ylabel(ax.get_ylabel(), fontsize=23, fontname='Arial')
# Adjust layout and save the figure
plt.tight_layout()
plt.savefig('deforested_data_corr_scatter.png', dpi=400)

plt.show()

g = sns.pairplot(forested_data_select_km, plot_kws={'s': 10, 'color': 'blue'}, diag_kws={'color': 'blue'})
for ax in g.axes.flatten():
    plt.setp(ax.get_xticklabels(), fontsize=12, fontname='Arial')
    plt.setp(ax.get_yticklabels(), fontsize=12, fontname='Arial')
    ax.set_xlabel(ax.get_xlabel(), fontsize=23, fontname='Arial')
    ax.set_ylabel(ax.get_ylabel(), fontsize=23, fontname='Arial')
# Adjust layout and save the figure
plt.tight_layout()
plt.savefig('forested_data_corr_scatter.png', dpi=400)

plt.show()
