==> scripts/python_operations/notebooks/3_plot_agreement_proportions.ipynb <==
This script plots the agreement proportions for the different years

import numpy as np
import matplotlib.pyplot as plt
import proplot as pplt

plt.rcParams.update({
    'xtick.labelsize': 20,
    'ytick.labelsize': 20,
    'legend.fontsize': 23,
    'axes.labelsize': 24,
    'axes.titlesize': 24
})

years = [2016, 2017, 2018, 2019, 2020]
categories = ['Low', 'Medium', 'High']
data_all = np.array([
    [0.4762686631, 0.1952759436, 0.3284553933],  # 2016
    [0.3833339066, 0.2892538587, 0.3274122347],  # 2017
    [0.4345898819, 0.3606897487, 0.2047203694],  # 2018
    [0.4371569394, 0.3599488802, 0.2028941805],  # 2019
    [0.4400152115, 0.3386557274, 0.2213290611]   # 2020
])

data_all = data_all * 100 #Convert to percent

# Stacked Bar Plot with simplified legend
fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(years))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(years, data_all[:, i], left=bottom, label=category, color=color)
    bottom += data_all[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Year',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_whole_madagascar.pdf", format='pdf', dpi=300)
plt.show()



# Years and categories
years = [2016, 2017, 2018, 2019, 2020]
categories = ['Low', 'Medium', 'High']

# Data for Spiny
data_spiny = np.array([
    [0.6495721649, 0.2098337249, 0.1405941102],  # 2016
    [0.4782901414, 0.3585688585, 0.1631410001],  # 2017
    [0.5543672499, 0.4173432840, 0.02828946613],  # 2018
    [0.5528840116, 0.4155191588, 0.03159682969],  # 2019
    [0.5529096741, 0.4191564275, 0.02793389843]   # 2020
])
data_spiny = data_spiny * 100
# Data for Dry
data_dry = np.array([
    [0.6408463089, 0.1480153912, 0.2111382999],  # 2016
    [0.5732683156, 0.2163711697, 0.2103605147],  # 2017
    [0.6080555405, 0.2729484739, 0.1189959856],  # 2018
    [0.6161037808, 0.2700029014, 0.1138933178],  # 2019
    [0.5497805173, 0.3042745480, 0.1459449347]   # 2020
])
data_dry = data_dry * 100
# Data for Moist
data_moist = np.array([
    [0.3310459216, 0.2248196996, 0.4441343788],  # 2016
    [0.2362706523, 0.3243013489, 0.4394279988],  # 2017
    [0.2944692542, 0.4108948391, 0.2946359067],  # 2018
    [0.2898275828, 0.4128492114, 0.2973232058],  # 2019
    [0.3593824881, 0.3389796026, 0.3016379092]   # 2020
])
data_moist = data_moist * 100
# Data for Mangroves
data_mangroves = np.array([
    [0.3361943183, 0.2155548607, 0.4482508209],  # 2016
    [0.3183022483, 0.3357602019, 0.3459375498],  # 2017
    [0.3261696474, 0.3516526841, 0.3221776685],  # 2018
    [0.3256839103, 0.3501388112, 0.3241772785],  # 2019
    [0.4102542615, 0.5648865943, 0.02485914417]  # 2020
])
data_mangroves = data_mangroves * 100

fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(years))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(years, data_spiny[:, i], left=bottom, label=category, color=color)
    bottom += data_spiny[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Year',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_spiny.pdf", format='pdf', dpi=300)
plt.show()

fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(years))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(years, data_dry[:, i], left=bottom, label=category, color=color)
    bottom += data_dry[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Year',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_dry.pdf", format='pdf', dpi=300)
plt.show()

fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(years))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(years, data_moist[:, i], left=bottom, label=category, color=color)
    bottom += data_moist[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Year',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_moist.pdf", format='pdf', dpi=300)
plt.show()

fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(years))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(years, data_mangroves[:, i], left=bottom, label=category, color=color)
    bottom += data_mangroves[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Year',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_mangroves.pdf", format='pdf', dpi=300)
plt.show()

average_spiny = np.mean(data_spiny, axis=0)
average_dry = np.mean(data_dry, axis=0)
average_moist = np.mean(data_moist, axis=0)
average_mangroves = np.mean(data_mangroves, axis=0)

# Combine all biomes by averaging across years
combined_data = np.array([average_spiny, average_dry, average_moist, average_mangroves])

# Resulting matrix: biome mean * proportion
combined_data

average_spiny

biomes = ['Spiny', 'Dry', 'Moist', 'Mangroves']
categories = ['Low', 'Medium', 'High']

# Plotting the data using proplot
fig, ax = pplt.subplots(figsize=(8, 6))
colors = ['red', 'yellow', 'green']
bottom = np.zeros(len(biomes))

for i, (category, color) in enumerate(zip(categories, colors)):
    ax.barh(biomes, combined_data[:, i], left=bottom, label=category, color=color)
    bottom += combined_data[:, i]

ax.format(
    xlabel='Proportion (%)',
    ylabel='Biome',
    xticklabelsize=18,
    yticklabelsize=18,
    xlabelsize=18,
    ylabelsize=18,
    titlesize=21,
)
ax.legend(bbox_to_anchor=(0, 1.15), loc='upper left', fontsize=30)

plt.savefig("agreement_proportions_mean_biomes.pdf", format='pdf', dpi=300)
plt.show()
