==> scripts/python_operations/notebooks/4_deforestation_rates.ipynb <==
import numpy as np
import geopandas as gpd


root = "/Users/dimeji/Documents/Research/2024/deforestation/final_project_version/data/tabular_data/"

data = gpd.read_file(root+"forested_area_biome_year_MFCFC.geojson")
data = data.drop(columns="id")
data["sum"] = data["sum"] / 1000000

data.head()

data.replace('Ecoregion Sud', 'spiny', inplace=True)
data.replace('Ecoregion Est', 'moist', inplace=True)
data.replace('Ecoregion Ouest', 'dry', inplace=True)

data2000 = data[data["year"] == 2000]

data2000

print("The total forested area in 2000 is: ", data2000["sum"].sum())

data2023 = data[data["year"] == 2023]

data2023

print("The total forested area in 2023 is: ", data2023["sum"].sum())

def annual_deforestation_rate(F_t2, F_t1, t2=2023, t1=2000):
    time_interval = t2 - t1
    rate = 100 * (1 - (1 - (F_t2 - F_t1) / F_t1) ** (1 / time_interval))
    return rate



print("Total def rate madagascar: ", abs(annual_deforestation_rate(data2023["sum"].sum(),  data2000["sum"].sum())))

print("Moist def rate madagascar: ", abs(annual_deforestation_rate(37418.70,  48555.82)))

print("Dry def rate madagascar: ", abs(annual_deforestation_rate(21461.18,  29414.22)))

print("Spiny def rate madagascar: ", abs(annual_deforestation_rate(16496.81,  18586.38)))

print("Mangroves def rate madagascar: ", abs(annual_deforestation_rate(1735.33,  1775.70)))

(0.10 / 100) * 1775.70
